Fix IncronTabEntry::GetSafePath
Previously only the space character (why?) and the backslash character
were escaped, leaving commands vulnerable to shell injection.

The correct way is to only escape the single-quote, and then put the path variables '$@' '$#' in them in commands. Other characters need not be escaped, as they will not be treated as special within single-quoted strings.